refactor: rename tool-classify.ts to normalizer-tool.ts
