+++ apps/api/src/engines/executors/claude/normalizer-tool.ts
